requirements_server.txt
# Requirements for OBD-II Real-Time Server (Raspberry Pi)
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.10.0
eventlet>=0.33.0
obd>=0.7.1
